fix(web-app): inputs #80
close #80
