+++ apps/web-app/src/models/UI/index.ts
@@ -1,4 +1,4 @@
-export * from './formData.model'
+export * from './inputs.model'
 export * from './theme.model'
 export * from './translation.model'
