--- apps/web-app/src/models/validations/index.ts
@@ -1,1 +1,0 @@
-export * from './user.validation'
