+++ apps/web-app/src/pages/Register/models/registerInputs.const.ts
@@ -1,27 +1,5 @@
-import { type TInputsFormData } from '@/models/UI'
-import { CUserValidations } from '@/models/validations'
-import { t } from 'i18next'
+import { CInputUser, type TInputsFormData } from '@/models/UI'
 
 export const inputsData: TInputsFormData = [
-  {
-    name: 'username',
-    label: t('register.form.main.inputs.userName'),
-    type: 'text',
-    validation: CUserValidations.username,
-    placeholder: 'Joe Bass'
-  },
-  {
-    name: 'email',
-    label: t('register.form.main.inputs.email'),
-    type: 'email',
-    validation: CUserValidations.email,
-    placeholder: 'joe@gmail.com'
-  },
-  {
-    name: 'password',
-    label: t('register.form.main.inputs.password'),
-    type: 'password',
-    validation: CUserValidations.password,
-    placeholder: 'wua-wau78'
-  }
+  CInputUser.username, CInputUser.email, CInputUser.password
 ]
